Moves lists to lua, improves list syntax.
The :ListThoughts and :ListThoughtTags have been moved over to the lua
implementation. Note that the python for these two has been removed.

The first part of the list is now highlighted with a region instead of
with three patterns. This makes the highlighting issue when there was a
'.' in the title go away.
Tags in the taglist are not /tag.: to take advantage of this change and
be highlighted appropriately.