Improves NewThought and adds SearchThoughtTags
This uses :enew instead of :e to create the buffer for a new tag. This
means that the filetype can be set after the template is inserted. This
should make it faster and more robust.

Removes code duplication in the thoughtbox#listThoughts functions.

Adds a new SearchThoughtTags which lists the tag first and then the rest
of the thought. This means each though is listed multiple times, but it
makes searching a bit more flexible.